Fix changes requested on code clean up
