+++ Decred Wallet/Extensions/UIColor.swift
@@ -16,6 +16,7 @@ extension UIColor {
         static let darkerGray = UIColor.init(hex: "#4e5f70")
         static let transparentThinGray = UIColor.init(hex: "#a4abb1", alpha: 0.3)
         static let shadowColor = UIColor(hex: "#3D0914")
+        static let lighterGrayGray = UIColor(hex: "#140000")
         
         // following are colors from new mockup color guide
         // ultimately other color constants should end up being obsolete once
