--- Decred Wallet/Extensions/UIViews.swift
@@ -1,30 +1,0 @@
-//
-//  UIView.swift
-//  Decred Wallet
-//
-// Copyright (c) 2018-2019 The Decred developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-import UIKit
-
-extension UIView {
-   func setRoundCorners(corners: UIRectCorner, radius: CGFloat) {
-        let path = UIBezierPath(roundedRect: bounds, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        let mask = CAShapeLayer()
-        mask.path = path.cgPath
-        layer.mask = mask
-    }
-    
-    func dropShadow(color: UIColor, opacity: Float = 0.2, offSet: CGSize, radius: CGFloat = 1, scale: Bool = true) {
-      layer.masksToBounds = false
-      layer.shadowColor = color.cgColor
-      layer.shadowOpacity = opacity
-      layer.shadowOffset = offSet
-      layer.shadowRadius = radius
-
-      layer.shadowPath = UIBezierPath(rect: self.bounds).cgPath
-      layer.shouldRasterize = true
-      layer.rasterizationScale = scale ? UIScreen.main.scale : 1
-    }
-}
