+++ Decred Wallet/Features/Wallet Setup/WalletSetup.storyboard
@@ -80,7 +80,7 @@
                                     </userDefinedRuntimeAttribute>
                                 </userDefinedRuntimeAttributes>
                                 <connections>
-                                    <action selector="RestoreWallet:" destination="98H-rf-Dma" eventType="touchUpInside" id="XqY-r0-F19"/>
+                                    <action selector="restoreWallet:" destination="98H-rf-Dma" eventType="touchUpInside" id="1Z5-EH-m3H"/>
                                 </connections>
                             </button>
                         </subviews>
